bots/gptbot/requirements.txt
python-dotenv==1.0.1
re-ircbot==2.0.15-dev
requests==2.32.3
